Better browser detection for tests...
We ask the browser itself for its navigator details.

From this info we can get a better idea of the O/S used, so that we can
name our test result files and tests with more browser-specific names.

webdriver reports WINDOWS for XP and Windows 7, and I want to be able to
see which O/S the tests were run on.